Document Scanner
The document scanner program implemented in C++ using the OpenCV library is a valuable addition to any portfolio showcasing image processing and computer vision skills. The program demonstrates the ability to preprocess images, detect contours, perform perspective transformation, and crop images to extract specific regions of interest. It showcases proficiency in working with the OpenCV library and highlights problem-solving skills in developing practical applications. This project is an excellent demonstration of the ability to apply computer vision techniques to real-world scenarios and provides evidence of the programmer's expertise in image processing algorithms and techniques.

Any photo document can be scanned with it.